Working on args
New branch with everything working in helpers so that I can start working on taking NPC names via new args.

Args appear to be working. Need to figure out how to deal with the fact that npcs MUST be combatants, but PCs may not be. Might require rewriting some helpers.